requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy>=2.0.0
alembic
asyncpg
psycopg2-binary
pgvector
pydantic
pydantic-settings
python-jose[cryptography]
passlib
# Fixed version to prevent compatibility issues
bcrypt==3.2.2
python-multipart
celery
redis
requests
pypdf
langchain-text-splitters
openai
google-generativeai
flask
flask-admin
flask-login
wtforms
httpx
# Required for async file uploads
aiofiles